updated Expander style
- updated expander style to match winui style
- common helpers moved to more appropriate paths
- SettingsExpander Expander animation is now fully in xaml and no longer uses Expander animation helper.
