+++ source/iNKORE.UI.WPF.Modern.Controls/Controls/Windows/FlipView/FlipView.cs
@@ -1500,63 +1500,4 @@ namespace iNKORE.UI.WPF.Modern.Controls
             UpdateVisualState(true);
         }
     }
-    
-    public class DoubleValueHolder : DependencyObject
-    {
-        public static readonly DependencyProperty XProperty =
-        DependencyProperty.Register("X", typeof(double), typeof(DoubleValueHolder), new PropertyMetadata(0.0));
-
-        public static readonly DependencyProperty YProperty =
-            DependencyProperty.Register("Y", typeof(double), typeof(DoubleValueHolder), new PropertyMetadata(0.0));
-
-        public static readonly DependencyProperty ZProperty =
-            DependencyProperty.Register("Z", typeof(double), typeof(DoubleValueHolder), new PropertyMetadata(0.0));
-
-        public double X
-        {
-            get { return (double)GetValue(XProperty); }
-            set { SetValue(XProperty, value); }
-        }
-
-        public double Y
-        {
-            get { return (double)GetValue(YProperty); }
-            set { SetValue(YProperty, value); }
-        }
-
-        public double Z
-        {
-            get { return (double)GetValue(ZProperty); }
-            set { SetValue(ZProperty, value); }
-        }
-    }
-
-    public class MultiplyConverter : IMultiValueConverter
-    {
-        public object Convert(object[] values, Type targetType, object parameter, CultureInfo culture)
-        {
-            var first = values[0];
-
-            if (first is not double result)
-            {
-                return values;
-            }
-
-            for (int index = 1; index < values.Length; index++)
-            {
-                if (values[index] is double current)
-                {
-                    result *= current;
-                }
-            }
-
-            return result;
-        }
-
-        public object[] ConvertBack(object value, Type[] targetTypes, object parameter, CultureInfo culture)
-        {
-            throw new NotImplementedException();
-        }
-    }
-    
 }
